# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend service for Leech-ru platform. Uses cookie-based authentication with HttpOnly tokens.",
        "title": "Leech API",
        "contact": {
            "name": "API Support",
            "email": "mmishin2107@gmail.com"
        },
        "version": "1.0"
    },
    "host": "пиявкипобеда.рф",
    "paths": {
...
        "/api/v1/main/{id}": {
            "get": {
                "description": "Retrieves a main page content using its UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "main"
                ],
                "summary": "Get content by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Content ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetByIdMainPageResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "404": {
                        "description": "Content not found",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Deletes the main page content with the given UUID.",
                "tags": [
                    "main"
                ],
                "summary": "Delete content by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Content ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted"
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "404": {
                        "description": "Content not found",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Updates main page content fields by given ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "main"
                ],
                "summary": "Update content",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Content ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated content fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMainPageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMainPageResponse"
                        }
                    },
                    "400": {
                        "description": "Validation or binding error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "404": {
                        "description": "Content not found",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
        "/api/v1/news": {
            "get": {
                "description": "Retrieves a list of news with pagination and content preview length option.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get news by filters",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Max number of items",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maximum": 500,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of characters for content preview",
                        "name": "preview_length",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.NewsListItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Creates a news article with provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Create a news article",
                "parameters": [
                    {
                        "description": "News data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateNewsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "409": {
                        "description": "Conflict: invalid news format",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/image/{image_id}": {
            "get": {
                "description": "Returns the image file stream by its UUID identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Download image by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image UUID",
                        "name": "image_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image file stream",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "404": {
                        "description": "Image not found",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateNewsRequest": {
            "type": "object",
            "required": [
                "content",
                "image_id",
                "is_hidden",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "example": "This is the full news content..."
                },
                "image_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "is_hidden": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Breaking News"
                }
            }
        },
        "dto.CreateNewsResponse": {
            "type": "object",
            "required": [
                "content",
                "id",
                "image_id",
                "is_hidden",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "example": "This is the full news content..."
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "image_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "is_hidden": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Breaking News"
                }
            }
        },
...
        "dto.GetByIdMainPageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "fluid": {
                    "type": "boolean"
                },
                "href": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "image_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174100"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.HTTPStatus": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.NewsListItem": {
            "type": "object",
            "required": [
                "content_preview",
                "id",
                "image_id",
                "is_hidden",
                "title"
            ],
            "properties": {
                "content_preview": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "This is a short preview..."
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "image_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "is_hidden": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Breaking News"
                }
            }
        },
...
        "dto.UpdateMainPageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 2
                },
                "fluid": {
                    "type": "boolean"
                },
                "href": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 2
                },
                "image_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "dto.UpdateMainPageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "fluid": {
                    "type": "boolean"
                },
                "href": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "image_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174100"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Authentication via HttpOnly cookies. System uses two cookies:\\n- `user_auth_access_token` (short-lived)\\n- `user_auth_refresh_token` (long-lived)\\n\\nAll protected endpoints require valid cookies to be automatically sent by browser.",
            "type": "apiKey",
            "name": "user_auth_access_token",
            "in": "cookie"
        }
    },
...
}